move functions to service and rename callback
